+++ finetuner/__init__.py
@@ -24,26 +24,31 @@ if HUBBLE_REGISTRY not in os.environ:
 
 from finetuner import callback, model
 from finetuner.console import print_model_table
+from finetuner.data import build_encoding_dataset
 from finetuner.experiment import Experiment
 from finetuner.finetuner import Finetuner
 from finetuner.model import list_model_classes
 
 if TYPE_CHECKING:
+    import numpy as np
     from _finetuner.models.inference import InferenceEngine
 
 ft = Finetuner()
 
 
-def login(force: bool = False):
+def login(force: bool = False) -> None:
+    """
+    Login to Jina AI to use cloud-based fine-tuning. Thereby, an authentication token is
+    generated which can be read with the :func:`~finetuner.get_token` function.
+
+    :param force: If set to true, an existing token will be overwritten. Otherwise,
+        you will not login again, if a valid token already exists.
+    """
     ft.login(force=force)
 
 
 def notebook_login(force: bool = False):
     ft.notebook_login(force=force)
-
-
-def connect():
-    ft.connect()
 
 
 def list_callbacks() -> Dict[str, callback.CallbackStubType]:
@@ -448,10 +453,10 @@ def get_model(
 
 
 def encode(
-    model,
-    data: DocumentArray,
+    model: 'InferenceEngine',
+    data: Union[DocumentArray, List[str]],
     batch_size: int = 32,
-):
+) -> Union[DocumentArray, 'np.ndarray']:
     """Preprocess, collate and encode the `DocumentArray` with embeddings.
 
     :param model: The model to be used to encode `DocumentArray`. In this case
@@ -469,6 +474,12 @@ def encode(
 
     from _finetuner.models.inference import ONNXRuntimeInferenceEngine
 
+    if isinstance(data, DocumentArray):
+        return_da = True
+    else:
+        data = build_encoding_dataset(model=model, data=data)
+        return_da = False
+
     for batch in data.batch(batch_size, show_progress=True):
         if isinstance(model, ONNXRuntimeInferenceEngine):
             inputs = model._run_data_pipeline(batch)
@@ -485,3 +496,5 @@ def encode(
             inputs = model._move_to_device(inputs)
             output = model.run(inputs)
             batch.embeddings = output.detach().cpu().numpy()
+
+    return data if return_da else data.embeddings