+++ CHANGELOG.md
@@ -10,17 +10,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add support for CSV files to the `fit` function. [#592](https://github.com/jina-ai/finetuner/pull/592)
+- Add support for CSV files to the `fit` function. ([#592](https://github.com/jina-ai/finetuner/pull/592))
+
+- Add support for lists to the `encode` function. [#598](https://github.com/jina-ai/finetuner/pull/598)
 
 ### Removed
 
+- Remove `connect` function. ([#596](https://github.com/jina-ai/finetuner/pull/596))
+
 ### Changed
 
+- Enhance documentation of login functionalities. ([#596](https://github.com/jina-ai/finetuner/pull/596))
+
 ### Fixed
+
+- Fix links to functions in the documentation. ([#596](https://github.com/jina-ai/finetuner/pull/596))
 
 ### Docs
 
 - Improve documentation on csv reading and run monitoring section. [#601](https://github.com/jina-ai/finetuner/pull/601)
+
+- Add documentation for `WandBLogger`. [#600](https://github.com/jina-ai/finetuner/pull/600)
 
 
 ## [0.6.4] - 2022-10-27